refactor: `MyEnvironmentPage` into two components
